move login into component, move routes to app.js
